# src/main/java/de/th/koeln/archilab/fae/faeteam2service/positionssender/PositionssenderApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.16).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package de.th.koeln.archilab.fae.faeteam2service.positionssender;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import java.util.List;

import javax.validation.Valid;
import javax.validation.constraints.NotNull;

import de.th.koeln.archilab.fae.faeteam2service.zone.ZoneDTO;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-01-23T20:04:16.757Z[GMT]")
@Api(value = "positionssender", description = "the positionssender API")
public interface PositionssenderApi {

    @ApiOperation(value = "Einen neuen Positionssender in die Datenbank aufnehmen", nickname = "addPositionssender", notes = "", response = PositionssenderDTO.class, tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 204, message = "Objekt wurde erstellt.", response = PositionssenderDTO.class),
            @ApiResponse(code = 405, message = "Invalide Daten.")})
    @RequestMapping(value = "/positionssender",
            produces = {"application/json"},
            consumes = {"application/json"},
            method = RequestMethod.POST)
    ResponseEntity<PositionssenderDTO> addPositionssender(@ApiParam(value = "Objekt, welches angelegt werden soll.", required = true) @Valid @RequestBody PositionssenderDTO body
    );


    @ApiOperation(value = "Liefert alle Positionssender.", nickname = "findAllPositionssender", notes = "", response = PositionssenderDTO.class, responseContainer = "List", tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "successful operation", response = PositionssenderDTO.class, responseContainer = "List"),
            @ApiResponse(code = 400, message = "ID ist nicht valide."),
            @ApiResponse(code = 404, message = "Positionssender konnte nicht gefunden werden.")})
    @RequestMapping(value = "/positionssender",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<List<PositionssenderDTO>> findAllPositionssender(@ApiParam(value = "Ergebnis nach Zone filtern") @Valid @RequestParam(value = "zoneId", required = false) String zoneId
    );


    @ApiOperation(value = "Liefert den Positionssendermit der ID", nickname = "findPositionssender", notes = "", response = PositionssenderDTO.class, tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "successful operation", response = PositionssenderDTO.class),
            @ApiResponse(code = 400, message = "ID ist nicht valide."),
            @ApiResponse(code = 404, message = "Positionssender konnte nicht gefunden werden.")})
    @RequestMapping(value = "/positionssender/{id}",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<PositionssenderDTO> findPositionssender(@ApiParam(value = "ID des Positionssender der zurückgeliefert werden soll.", required = true) @PathVariable("id") String id
    );


    @ApiOperation(value = "Liefert eine Liste aller Positionssendermit der ID", nickname = "findPositionssenderInRadius", notes = "", response = PositionssenderDTO.class, responseContainer = "List", tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Liste aller Positionssender im Radius", response = PositionssenderDTO.class, responseContainer = "List"),
            @ApiResponse(code = 405, message = "Invalide Daten")})
    @RequestMapping(value = "/positionssender/findInRadius",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<List<PositionssenderDTO>> findPositionssenderInRadius(@NotNull @ApiParam(value = "Radius um die Position, in welcher gesucht werden soll.", required = true) @Valid @RequestParam(value = "radius", required = true) Double radius
            , @NotNull @ApiParam(value = "Laengengrad des Radius um die Position, in welcher gesucht werden soll.", required = true) @Valid @RequestParam(value = "laengengrad", required = true) Double laengengrad
            , @NotNull @ApiParam(value = "Breitengrad des Radius um die Position, in welcher gesucht werden soll.", required = true) @Valid @RequestParam(value = "breitengrad", required = true) Double breitengrad
    );


    @ApiOperation(value = "Liefer alle Zonen des Positionssenders", nickname = "getZonenByPositionssenderId", notes = "Liefert alle Zonene des Positionssenders", response = ZoneDTO.class, responseContainer = "List", tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "successful operation", response = ZoneDTO.class, responseContainer = "List"),
            @ApiResponse(code = 400, message = "Invalid ID supplied"),
            @ApiResponse(code = 404, message = "Pet not found")})
    @RequestMapping(value = "/positionssender/{id}/zone",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<List<ZoneDTO>> getZonenByPositionssenderId(@ApiParam(value = "ID des Positionssenders", required = true) @PathVariable("id") String id
    );


    @ApiOperation(value = "Daten eines Positionssenders aktualisieren", nickname = "updatePositionssender", notes = "", response = PositionssenderDTO.class, tags = {"Positionssender"})
    @ApiResponses(value = {
            @ApiResponse(code = 200, message = "Objekt wurde aktualisiert.", response = PositionssenderDTO.class),
            @ApiResponse(code = 400, message = "Die ID ist nicht valide."),
            @ApiResponse(code = 404, message = "Objekt konnte nicht gefunden werden."),
            @ApiResponse(code = 405, message = "Validationsfehler")})
    @RequestMapping(value = "/positionssender/{id}",
            produces = {"application/json"},
            consumes = {"application/json"},
            method = RequestMethod.PUT)
    ResponseEntity<PositionssenderDTO> updatePositionssender(@ApiParam(value = "Objekt eines Positionssenders, welches aktualisiert werden soll.", required = true) @Valid @RequestBody PositionssenderDTO body
            , @ApiParam(value = "ID des Positionssender der aktualisiert werden soll.", required = true) @PathVariable("id") String id
    );

}